refactor: API, 서비스 위치 Event -> EventParticipation 으로 이동
